Refine resumable session helpers
